// generated/prisma/models/GearItem.ts
export type GearItemWhereInput = {
  AND?: Prisma.GearItemWhereInput | Prisma.GearItemWhereInput[]
  OR?: Prisma.GearItemWhereInput[]
  NOT?: Prisma.GearItemWhereInput | Prisma.GearItemWhereInput[]
  id?: Prisma.StringFilter<"GearItem"> | string
  name?: Prisma.StringFilter<"GearItem"> | string
  description?: Prisma.StringFilter<"GearItem"> | string
  brand?: Prisma.StringNullableFilter<"GearItem"> | string | null
  pricePerDay?: Prisma.FloatFilter<"GearItem"> | number
  stock?: Prisma.IntFilter<"GearItem"> | number
  isAvailable?: Prisma.BoolFilter<"GearItem"> | boolean
  images?: Prisma.StringNullableListFilter<"GearItem">
  categoryId?: Prisma.StringFilter<"GearItem"> | string
  providerId?: Prisma.StringFilter<"GearItem"> | string
  createdAt?: Prisma.DateTimeFilter<"GearItem"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"GearItem"> | Date | string
  category?: Prisma.XOR<Prisma.CategoryScalarRelationFilter, Prisma.CategoryWhereInput>
  provider?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
  reviews?: Prisma.ReviewListRelationFilter
  rentalOrders?: Prisma.RentalOrderListRelationFilter
}
// ...
export type StringNullableListFilter<$PrismaModel = never> = {
  equals?: string[] | Prisma.ListStringFieldRefInput<$PrismaModel> | null
  has?: string | Prisma.StringFieldRefInput<$PrismaModel> | null
  hasEvery?: string[] | Prisma.ListStringFieldRefInput<$PrismaModel>
  hasSome?: string[] | Prisma.ListStringFieldRefInput<$PrismaModel>
  isEmpty?: boolean
}